+++ lib/generators/upload_assets_generator/install/upload_assets_templates/assets/stylesheets/components/direct_uploads.sass
