+++ lib/generators/upload_assets_generator/install/upload_assets_templates/javascript/controllers/direct_uploader_controller.js
