+++ lib/generators/upload_assets_generator/install/upload_assets_templates/javascript/packs/custom_uploader.js
