+++ lib/generators/upload_assets_generator/install/upload_assets_templates/javascript/packs/direct_uploads.js
