+++ lib/video_upload_field/inputs.rb
@@ -1,0 +1,6 @@
+module VideoUploadField
+  module Inputs
+    extend ActiveSupport::Autoload
+    autoload :FileUploadInput
+  end
+end
