+++ video_upload_field.gemspec
@@ -27,6 +27,9 @@ Gem::Specification.new do |spec|
   spec.bindir        = "exe"
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib"]
+  spec.add_dependency 'rails', '~> 6.0.3'
+  spec.add_dependency 'simple_form', '~> 5.0.2'
+
 
   spec.add_development_dependency "bundler", "~> 2.0"
   spec.add_development_dependency "rake", "~> 10.0"
